requirements.txt
numpy>=1.23
opencv-python>=4.7
matplotlib>=3.7
torch>=2.0
torchvision>=0.15
pytest>=7.0
ultralytics>=8.0.0
pyyaml>=6.0
